修复： 1.同时触发两种不同种类的陷阱时的情况 [增加了判定变量，一帧只能触发一次] 2.遇到写Debug.Log逻辑就恢复正常的话，记得查看是否是.dll的缓存问题，可以清除一下Library
新增：
1.更新了格挡时的表现
2.拓展了MayHurt函数
3.基本完成了和投掷物相关的事物

待办：
1.滚动列表的真正实现

+++ Assets/Script/Enemy/Common/Enemy.cs
@@ -5,7 +5,7 @@ namespace MyEnemy
     public partial class Enemy : MonoBehaviour
     {
         public virtual void Death(){}
-        public virtual void Hurt(float damage){}
+        public virtual void Hurt(int damage){}
     }
 }
